requirements.txt
flask
flask-restx
flask-cors
python-dotenv
google-generativeai
google.genai
marshmallow
flasgger